# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Colab Server API
  version: 0.1.4-draft
  summary: Colab 서버 REST API — P1~P3(수직 슬라이스 · 협업 코어 · HITL/재개/예산)에서 웹(W)과 colab CLI(C)가 호출하는 전부.
  description: |
    # 읽는 법

    - **SSOT**: 리소스·필드·ENUM 값은 `server/migrations/0001_init.sql`(스키마 v0)과 정확히 일치한다. 동작은 `PRD.md` FR-1~FR-9, 화면이 부르는 순서는 `SCREEN.md` §4.
    - **권한**은 각 operation `description`의 **첫 줄**에 `권한:`으로 적는다. 원칙(PRD §9): 판정은 항상 체인 최상단의 **사람 originator** 기준이며 에이전트 경유로 권한이 상승하지 않는다. 세션 내 사람 역할은 Director · deputy · 워크스페이스 멤버 셋뿐이다(FR-5.3).
    - **데몬 엔드포인트**(claim · heartbeat · event push · 토큰 폐기 통보 · 페어링 토큰 교환)는 이 문서에 없다 — `contracts/daemon-protocol.md`.
    - **colab CLI가 쓰는 operation**은 `x-colab-cli: <명령>` 확장으로 표시하고 보안 스킴 `TaskToken`(`COLAB_TASK_TOKEN`, task 범위·단기)을 쓴다. 토큰이 재큐잉으로 폐기되면 `401`이다(FR-9.1).
    - **오류**는 전부 RFC 9457 `application/problem+json`(`Problem`). 검증 실패는 `422`, 상태 전이 불가·중복·차단은 `409`, 권한 없음은 `403`(버튼은 보이되 비활성 — 사유는 `detail`).
    - **페이지네이션**은 커서 방식(`cursor` · `limit` → `next_cursor`). 메시지 타임라인만 `before`/`after` 양방향.
    - **멱등키** `Idempotency-Key`(UUID)는 메시지 게시(`POST /sessions/{id}/messages`, `POST /lanes/{id}/restart`)와 HITL 응답(`POST /hitl-requests/{id}/response`)에 **필수**, 그 밖의 생성 POST에서 선택. 같은 키의 재요청은 첫 응답을 그대로 돌려준다(`Idempotent-Replayed: true`).
    - **실시간**은 SSE 하나(`GET /workspaces/{workspaceId}/stream`). 이벤트 종류는 `StreamEvent.type`에 전부 나열한다.
    - `x-prd`: 그 operation이 구현하는 PRD 항목. `x-screen`: 부르는 화면. `x-phase`: 처음 필요한 단계(P1~P5).
  license:
    name: Proprietary
...
servers:
...
security:
  - UserSession: []
...
paths:
...
  /inbox:
    get:
      tags: [inbox]
      operationId: listInbox
      summary: 인박스(S8)
      description: |
        권한: 로그인한 사용자(본인 멤버십의 항목만).
        정의는 "나에게 열려 있는 HITL 요청 + 기한 임박 항목"의 합집합(FR-5.4). **정렬**: `overdue` → `action_required` → `attention` → `info`, 묶음 안에서 기한 임박 순(SCREEN §4.6). `hitl_request`는 처음엔 Director에게만, 기한 절반 후 deputy에게도 나타나며 `delegated: true`(O5). 각 항목은 타입별 카드 데이터(`ref`)와 인라인 동작 목록(`actions`)을 가진다.
      x-prd: [FR-8, FR-5.4, FR-6.2.1]
      x-screen: [S8]
      x-phase: P2
      parameters:
        - name: workspace_id
          in: query
          description: 비우면 모든 워크스페이스.
          schema: { type: string, format: uuid }
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, unread, action_required]
            default: all
        - name: session_id
          in: query
          schema: { type: string, format: uuid }
        - name: type
          in: query
          style: form
          explode: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/InboxItemType' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 페이지(정렬 규칙 적용).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/InboxItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /inbox/summary:
    get:
      tags: [inbox]
      operationId: getInboxSummary
      summary: 내비 뱃지 — action_required 개수만
      description: |
        권한: 로그인한 사용자.
        `info`까지 세면 뱃지가 무의미하다(SCREEN §4.6).
      x-screen: [App Nav]
      x-phase: P2
      parameters:
        - name: workspace_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 개수.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboxSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /inbox/{inboxItemId}/read:
    parameters:
      - name: inboxItemId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [inbox]
      operationId: markInboxRead
      summary: 읽음 처리
      description: |
        권한: 항목 소유 멤버.
        멱등. `action_required` 항목은 읽어도 해소되지 않는다 — 해소는 응답/답글이 한다.
      x-screen: [S8]
      x-phase: P2
      responses:
        '200':
          description: 읽음.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboxItem' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /inbox/read-all:
    post:
      tags: [inbox]
      operationId: markAllInboxRead
      summary: 전부 읽음(info · attention만)
      description: |
        권한: 로그인한 사용자.
        `action_required`는 건드리지 않는다.
      x-screen: [S8]
      x-phase: P2
      parameters:
        - name: workspace_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 처리 수.
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── artifacts · decisions ─────────────────────────────
...
components:
...
  securitySchemes:
    UserSession:
      type: apiKey
      in: cookie
      name: colab_session
      description: 로그인(S1/S2)으로 발급되는 사용자 세션 쿠키(HttpOnly · SameSite=Lax). 웹 클라이언트 기본 인증. 사람의 모든 권한 판정은 이 사용자 = originator 기준.
    TaskToken:
      type: http
      scheme: bearer
      bearerFormat: COLAB_TASK_TOKEN
      description: |
        task마다 발급되는 단기 토큰(`COLAB_TASK_TOKEN`, 형식은 `contracts/colab-cli.md`). 그 task의 세션·에이전트 범위 밖에는 접근할 수 없다(FR-7.4). 재큐잉·취소 시 서버가 폐기하며(FR-9.1) 이후 호출은 `401 token_revoked`. 테스트 채팅에는 발급되지 않는다(FR-1.8.1). `x-colab-cli`가 붙은 operation만 이 스킴을 받는다.
    DaemonToken:
      type: http
      scheme: bearer
      bearerFormat: COLAB_DAEMON_TOKEN
      description: |
        페어링(`daemon-protocol.md` §2)으로 런타임마다 발급되는 데몬 토큰(`cdt_…`). 데몬 API(`/v1/daemon/*`)는 이 문서가 아니라 `daemon-protocol.md`가 스펙이므로 여기 등장하는 operation은 **`downloadArtifact` 하나**다(v0.7.3, T-I4): §4.3 `rebind_prepare`가 데몬에게 아티팩트 다운로드를 지시하기 때문이다. 범위는 **그 런타임에 고정된 세션**의 아티팩트뿐이고, 런타임 삭제·토큰 회수로 폐기된다.

  # ───────────────────────────── parameters ─────────────────────────────
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: 이전 응답의 `next_cursor`. 불투명 문자열.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    Problem:
      description: 그 밖의 오류. 본문은 항상 `application/problem+json`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: "인증 없음 · 사용자 세션 만료 · TaskToken 폐기(`code: token_revoked`) 또는 만료."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            revoked:
              value: { type: 'https://colab.dev/problems/token-revoked', title: Task token revoked, status: 401, code: token_revoked, detail: 이 task는 재큐잉되어 토큰이 폐기되었다. 즉시 종료하라. }
...
    NotFound:
      description: 없음(또는 다른 워크스페이스의 리소스 — 존재를 노출하지 않는다).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details + `code`(기계용 식별자) + `errors[]`(필드 검증) 확장.
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: 사람에게 보여줄 사유(화면의 비활성 툴팁·경고 문구). }
        instance: { type: string, format: uri }
        code:
          type: string
          description: 기계용 식별자(snake_case). 예 — `token_revoked` · `deputy_not_yet` · `hitl_already_open` · `not_participant` · `runtime_offline` · `no_runtime` · `workdir_dirty` · `runtime_has_active_sessions` · `invalid_transition` · `idempotency_key_reused` · `validation_failed` · `not_designated_reviewer` · `workdir_quota_exceeded` · `running_lanes`.
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, description: JSON pointer 또는 필드명. }
              code: { type: string }
              message: { type: string }
        can_respond_from:
          type: [string, 'null']
          format: date-time
          description: "deputy 의 시점 제한(기한 절반 경과 시각)일 때만 시각. **권한이 영영 없는 사람(일반 멤버 등)에게는 비운다(null)** — 생기지 않을 권리에 시각을 약속하지 않는다(E7-11, PR #108)."
        sessions:
          type: array
          description: 차단 사유가 된 세션(런타임 삭제 409 등).
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              title: { type: string }
        workdirs:
          type: array
          description: 차단 사유가 된 작업 폴더(deleteSession 409 `workdir_unmerged`).
          items: { $ref: '#/components/schemas/Workdir' }
      additionalProperties: true
    Page:
      type: object
      description: 커서 페이지 공통 봉투. `items`는 각 operation이 정한다.
      required: [items, next_cursor]
      properties:
        items: { type: array, items: {} }
        next_cursor:
          type: [string, 'null']
          description: 다음 페이지 커서. null이면 끝.

    # ── ENUM (0001_init.sql과 동일) ──
...
    SessionStatus:
      type: string
      enum: [draft, active, paused, completing, completed, cancelled]
      description: '`session_status` (FR-2.3)'
    PauseReason:
      type: string
      enum: [budget, time, loop, runtime_offline, director]
      description: '`pause_reason` (FR-2.3 · FR-7.3 · FR-9.2)'
...
    WorkdirKind:
      type: string
      enum: [worktree, container, dir]
      description: '`workdir_kind`'
    WorkdirStatus:
      type: string
      enum: [active, retained, deleted]
      description: '`workdir_status`'
...
    FailureKind:
      type: string
      enum: [auth, quota, config, network, runtime_offline, stall, timeout, cancelled, other]
      description: '`failure_kind`. auth · quota · config는 재시도 없음(→ 에이전트 `error`).'
...
    HitlType:
      type: string
      enum: [question, choice, approval, info]
      description: '`hitl_type` (FR-5.1). v1은 question · approval.'
...
    InboxItemType:
      type: string
      enum: [hitl_request, lane_blocked, session_completed, session_paused, run_failed, runtime_offline, mention, workdir_gc_blocked]
      description: '`inbox_item_type` (FR-8). `workdir_gc_blocked`(P4, FR-6.4 M4 · E13-12·13) — 보존 기한이 지난 worktree 를 미병합 커밋·미커밋 변경 때문에 지우지 못했다. 같은 workdir 에 미해결 항목이 있으면 다시 만들지 않는다(스윕 멱등). 카드: {workdir_id, session_id, repo_path, branch, reason, commits_ahead}'
    InboxSeverity:
      type: string
      enum: [action_required, attention, info]
      description: '`inbox_severity` (SCREEN §4.6)'
...
    Workdir:
      type: object
      required: [id, session_id, kind, path_or_ref, status, disk_bytes, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        session: { $ref: '#/components/schemas/SessionRef' }
        agent_id: { type: [string, 'null'], format: uuid, description: worktree — 에이전트당 1개. }
        lane_id: { type: [string, 'null'], format: uuid, description: container · none — lane당 1개. }
        kind: { $ref: '#/components/schemas/WorkdirKind' }
        path_or_ref: { type: string }
        branch: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/WorkdirStatus' }
        disk_bytes: { type: integer, format: int64 }
        last_used_at: { type: [string, 'null'], format: date-time }
        retain_until: { type: [string, 'null'], format: date-time }
        dirty:
          type: [boolean, 'null']
          description: 미병합 커밋 또는 미커밋 변경이 있음(GC 차단 · 수동 삭제 경고, FR-6.4 M4). 데몬 마지막 보고 기준.
        merged:
          type: [boolean, 'null']
          description: "P4(T-S9 질문 2): 브랜치가 기본 브랜치에 병합됐는가. 데몬 마지막 보고 기준."
        commits_ahead:
          type: [integer, 'null']
          description: "P4: 기본 브랜치 대비 이 브랜치의 커밋 수. 0 이면 E13-11(커밋 0 + 클린 = 삭제)."
        gc_blocked_reason:
          type: [string, 'null']
          enum: [unmerged_commits, uncommitted_changes, null]
          description: "P4: 마지막 GC 판정이 삭제를 막은 사유. Director 의 다음 행동이 다르다 — `unmerged_commits` 는 병합해라(E13-12), `uncommitted_changes` 는 커밋하거나 버려라(E13-13). null = 차단 없음. deleteWorkdir 409 의 Problem.detail 에도 같은 값."
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ── 에이전트 ──
...
    SessionRef:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: '#/components/schemas/SessionStatus' }
...
    InboxItem:
      type: object
      required: [id, workspace_id, type, severity, session_id, ref_id, read_at, created_at, actions]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/InboxItemType' }
        severity: { $ref: '#/components/schemas/InboxSeverity' }
        session_id: { type: [string, 'null'], format: uuid }
        session: { $ref: '#/components/schemas/SessionRef' }
        ref_id: { type: [string, 'null'], format: uuid, description: 타입별 — hitl_request · message · task · runtime id. }
        due_at: { type: [string, 'null'], format: date-time, description: hitl_request면 기한(정렬 키). }
        overdue: { type: boolean }
        delegated: { type: boolean, description: 'deputy에게 위임되어 나타난 항목 — "위임됨 · 지금부터 응답 가능"(O5).' }
        card:
          type: object
          description: 타입별 카드 내용(SCREEN §4.6 표). 세션을 열지 않고 처리할 수 있게 충분히 담는다.
          properties:
            title: { type: string }
            body: { type: string }
            agent_name: { type: [string, 'null'] }
            proposed_default: { type: [string, 'null'] }
            hitl_type: { $ref: '#/components/schemas/HitlType' }
            paused_reason: { $ref: '#/components/schemas/PauseReason' }
            purpose: { type: [string, 'null'], enum: [agent, user_approval, budget, time, loop, null], description: "HITL 항목이면 `HitlRequest.purpose` 를 그대로 싣는다(K-9) — 웹이 task 범위 예산 HITL(세션은 active)에 상향 입력을 붙이려면 카드에서 바로 읽어야 한다. 비-HITL 항목은 null." }
            failure_kind: { $ref: '#/components/schemas/FailureKind' }
            runtime_name: { type: [string, 'null'] }
            grace_ends_at: { type: [string, 'null'], format: date-time }
            message_id: { type: [string, 'null'], format: uuid }
            lane_id: { type: [string, 'null'], format: uuid }
            summary: { type: [string, 'null'], description: session_completed 결과 요약. }
        actions:
          type: array
          description: 인라인 동작(타입·권한별).
          items: { type: string, enum: [answer, approve, reject, reply, approve_continue, restart, rebind, open_session, open_runtimes] }
        read_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    InboxSummary:
      type: object
      required: [action_required, overdue]
      properties:
        action_required: { type: integer, description: 내비 뱃지 값. }
        overdue: { type: integer }
        unread: { type: integer }

    # ── 비용 ──
...